[Fix]: Address the requested changes
