+++ utils/yolov5/general.py
@@ -34,7 +34,7 @@ import torch
 import torchvision
 import yaml
 
-from utils import TryExcept, emojis
+from . import TryExcept, emojis
 # from utils.downloads import gsutil_getsize
 # from utils.metrics import box_iou, fitness
